refactor(client): consolidate dead settings feature dirs into settings-* pages
Move 6 standalone feature components into their consolidated settings page directories,
eliminating redundant top-level feature directories that were only accessible via the
settings-access and settings-security container components.

- allowlist, github-allowlist, repo-blocklist → settings-access/
- security-overview, wallets, spending → settings-security/

Update imports in both consolidated page components. All routes and redirects unchanged.
Also fix pre-existing import order lint error in server/routes/settings.ts.
