+++ client/src/app/features/settings-security/wallet-viewer.component.ts
